Moved generateJSON method from SwaggerSpecJsonGenerator to the OpenAPISpecModel class.
